supplementary fix for 3410ea8945d66594fcd5e9a071b0a460fdc5b324

+++ source/core/mining-manager/faceted-submanager/NumericRangeGroupCounter.cc
@@ -27,10 +27,10 @@ NumericRangeGroupCounter* NumericRangeGroupCounter::clone() const
 
 void NumericRangeGroupCounter::addDoc(docid_t doc)
 {
-    int value;
+    int64_t value;
     if (propertyTable_->convertPropertyValue(doc, value))
     {
-        value = std::min(value, bound_[LEVEL_1_OF_SEGMENT_TREE] - 1);
+        value = std::min(value, int64_t(bound_[LEVEL_1_OF_SEGMENT_TREE] - 1));
         int exponent;
         for (exponent = 1; value >= bound_[exponent]; ++exponent);
         value /= bound_[exponent--] / 100;
